main_weather_app/requirements.txt
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.6
djangorestframework==3.14.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
psycopg2==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
